requirements.txt
Django==4.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
whitenoise==6.6.0
Pillow==10.0.1
django-admin-interface==0.24.0
django-colorfield==0.8.0
sqlparse==0.4.4
reportlab==4.0.4
python-dateutil==2.9.0.post0